Fixed sim_eliminate_resource_smart quest never activating
In the original game, when the point to capture is target_resource_smart ('resource' or 'territory' points), the game is supposed to give the player a different quest that sim_eliminate_smart. It works fine with sim_eliminate_path_smart (which has a different name - "Clear the road" - and description - "Help the friendly squad get through"), however, for whatever reason, it does not work with sim_eliminate_resource_smart (which has the same name but a different description - "Chase the enemy away from the area. This will weaken him.")

This commit fixes that issue.